+++ package.js
@@ -138,9 +138,9 @@ const specs = {
   },
 
   peerDependencies: {
-    react:                            '^15.3.0',
-    'react-dom':                      '^15.3.0',
-    moment: '^2.0.0',
+    react: '^15.3.0',
+    'react-dom': '^15.3.0',
+    moment: '^2.0.0',         // optional
   },
 
   dependencies: {