+++ src/inputs/timePickerDigital.js
@@ -1,5 +1,5 @@
 import React                from 'react';
-import moment               from 'moment';
+import moment               from '../vendor/moment';
 import { bindAll }          from '../gral/helpers';
 import {
   startOfToday,